[update] targetWeatherTimeList をリアルタイムで更新できるように変更
targetWeather => targetWeather以外 で targetWeatherTimeList に currentWeatherTime を追加するのではなく
targetWeather => targetWeather    で targetWeatherTimeList の currentWeatherTime を更新
targetWeather以外 => targetWeather で targetWeatherTimeList に currentWeatherTime を追加

+++ src/main/WeatherValue.java
@@ -12,7 +12,7 @@ public class WeatherValue {
     public List<String> pastWeatherList;
     public List<Integer> targetWeatherTimeList;
     public int totalTargetWeatherTime = 0;
-    public Long totalTargetWeatherCount = (long) 0;
+    public int totalTargetWeatherCount = 0;
 
     WeatherValue(String city, String targetWeather, String currentWeather, List<String> pastWeatherList) {
         this.city = city;
